ml-service/requirements.txt
Flask==3.0.3
pymongo==4.8.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
python-dotenv==1.0.1
